Move list proxy into its own file
