+++ tests/test_collections.py
@@ -1,8 +1,9 @@
 from unittest.mock import Mock
 
 from observ.dep import Dep
-from observ.observables import dict_traps, list_traps
-from observ.observables import DictProxy, ListProxy
+from observ.list_proxy import list_traps, ListProxy
+from observ.observables import dict_traps
+from observ.observables import DictProxy
 from observ.observables import proxy_db
 from observ.set_proxy import set_traps, SetProxy
